code cleanup, working and feature complete version

+++ PyDedupFS.py
@@ -1,4 +1,4 @@
-#/usr/bin/env python
+#!/usr/bin/env python
 # -*- coding: utf-8 -*- 
 #
 # This program is free software; you can redistribute it and/or modify
@@ -36,34 +36,50 @@ fuse.fuse_python_api = (0, 2)
 fuse.feature_assert('stateful_files', 'has_init')
 # Logging
 import logging
-logging.basicConfig(level=logging.DEBUG)
-# Profiling
-import cProfile
-import pstats
+import logging.config
+# read logging config from logging.conf
+logging.config.fileConfig("logging.conf")
+# hashing library
+import hashlib
 # own modules
 from MetaStorage import MetaStorage as MetaStorage
-from PyDedupFSExceptions import *
 from PyDedupFile import PyDedupFile as PyDedupFile
+
+# DEFAULT values use, if no others are applied
+DEFAULT_BLOCKSIZE = 1024 * 128
+DEFAULT_HASHFUNC = "sha1"
 
 class PyDedupFS(fuse.Fuse):
     """Fuse Interface Class"""
 
     def __init__(self, *args, **kw):
-        self.logger = logging.getLogger()
-        self.logger.setLevel(logging.INFO)
+        self.logger = logging.getLogger(self.__class__.__name__)
         self.logger.debug("PyDedupFS.__init__(%s, %s)", args, kw)
         # TODO : are there other parameters to set
         self.direct_io = False
         self.keep_cache = False
-        # dirty flag
+        self.fsname = "PyDedupFS"
+        # define meta_storage
+        self.meta_storage = None
+        # the last element of command line options
+        # is the mount_point
+        self.mountpoint = sys.argv[-1]
+        # dirty flag, to indicate, some data must be written to disk
         self.isdirty = False
         fuse.Fuse.__init__(self, *args, **kw)
+        # set default values, 
+        # cmd options have to be initialized in __init__ first
+        self.base = None
+        self.blocksize = DEFAULT_BLOCKSIZE
+        self.str_hashfunc = "sha1"
+        self.hashfunc = hashlib.sha1
         # wrapper to own file class, to pass meta_storage object
-        class wrapped_file_class(PyDedupFile):
+        # has to be in __init__ to function properly
+        class PyDedupFileWrapper(PyDedupFile):
             def __init__(self2, *a, **kw):
                 PyDedupFile.__init__(self2, self.meta_storage, *a, **kw)
         # set own file class
-        self.file_class = wrapped_file_class
+        self.file_class = PyDedupFileWrapper
 
     def getattr(self, path):
         """
@@ -71,10 +87,7 @@ class PyDedupFS(fuse.Fuse):
         return errno.ENOENT if File does not exist
         """
         self.logger.debug("PyDedupFS.getattr(%s)", path)
-        try:
-            return(self.meta_storage.getattr(path))
-        except NoEntry:
-            return(-fuse.ENOENT)
+        return(self.meta_storage.getattr(path))
 
     def readlink(self, path):
         """
@@ -82,8 +95,7 @@ class PyDedupFS(fuse.Fuse):
         NOT IMPLEMENTED : it doesnt make sense on this type of filesystem
         """
         self.logger.debug("PyDedupFS.readlink(%s)", path)
-        return(0)
-        
+        return(None)
 
     def readdir(self, path, offset):
         """
@@ -109,17 +121,23 @@ class PyDedupFS(fuse.Fuse):
     def truncate(self, path, offset):
         """change the size of a file"""
         self.logger.debug("PyDedupFS.truncate(%s, offset=%s)", path, offset)
-        return(0)
 
     def symlink(self, path, symlink):
         """
         create symlink from path to path1
         path must exist, symlink must not exist
         path is no absolute path, relative to actual position
-        NOT IMPLEMENTED : it doesnt make sense on this type of filesystem
+        NOT IMPLEMENTED, conecpt of symlinks is realy hard to implement in concep of PyDedupFS
         """
         self.logger.debug("PyDedupFS.symlink(%s, %s)", path, symlink)
-        return(0)
+        # path is the real existing file
+        # given relative to fuse mountpoint
+        #relpath = os.path.relpath(path, self.mountpoint)
+        #self.logger.debug("corrected relpath:%s", os.path.relpath(path, self.mountpoint))
+        #self.logger.debug("corrected absolute path:%s", os.path.normpath(os.path.join(self.mountpoint, relpath)))
+        #self.logger.debug("corrected absolute path:%s", os.path.normpath(os.path.join(self.mountpoint, path)))
+        #self.meta_storage.symlink(path, symlink)
+        return(None)
 
     def rename(self, path, path1):
         """
@@ -144,6 +162,8 @@ class PyDedupFS(fuse.Fuse):
         # inode numbers?).
         """
         self.logger.debug("PyDedupFS.link(%s, %s)", path, path1)
+        # map link to copy
+        self.meta_storage.copy(path, path1)
 
     def chmod(self, path, mode):
         """
@@ -168,10 +188,11 @@ class PyDedupFS(fuse.Fuse):
         we dont want special Files
         NOT IMPLEMENTED : we dont want special Files
         """
-        self.logger.debug("PyDedupFS.mknod(%s, mode=%s, dev=%s)", path, mode, dev) 
-        return(0)
+        self.logger.debug("PyDedupFS.mknod(%s, mode=%s, dev=%s)", path, mode, dev)
+        return(None)
 
     def mkdir(self, path, mode):
+        """create directory"""
         self.logger.debug("PyDedupFS.mkdir(%s, mode=%s)", path, mode)
         self.meta_storage.mkdir(path, mode)
         return(0)
@@ -197,7 +218,6 @@ class PyDedupFS(fuse.Fuse):
         -errno.ENOENT if file does not exists or other error
         """
         self.logger.debug("PyDedupFS.access(%s, mode=%s)", path, mode)
-        return(0)
 
     def statfs(self):
         """
@@ -218,18 +238,17 @@ class PyDedupFS(fuse.Fuse):
             - f_ffree - nunber of free file inodes
         """
         self.logger.debug("PyDedupFS.statfs()")
-        blocksize = 1024 * 128
         # from https://github.com/xolox/dedupfs/blob/master/dedupfs.py
         # TODO make this a parameter
-        host_fs = os.statvfs("/home/mesznera/pydedupfs")
+        host_fs = os.statvfs(self.base)
         return fuse.StatVfs(
-            f_bavail = (host_fs.f_bsize * host_fs.f_bavail) / blocksize, 
+            f_bavail = (host_fs.f_bsize * host_fs.f_bavail) / self.blocksize, 
             # The total number of free blocks available to a non privileged process.
-            f_bfree = (host_fs.f_frsize * host_fs.f_bfree) / blocksize, 
+            f_bfree = (host_fs.f_frsize * host_fs.f_bfree) / self.blocksize, 
             # The total number of free blocks in the file system.
-            f_blocks = (host_fs.f_frsize * host_fs.f_blocks) / blocksize, 
+            f_blocks = (host_fs.f_frsize * host_fs.f_blocks) / self.blocksize, 
             # The total number of blocks in the file system in terms of f_frsize.
-            f_bsize = blocksize, 
+            f_bsize = self.blocksize, 
             # The file system block size in bytes.
             f_favail = 0, 
             # The number of free file serial numbers available to a non privileged process.
@@ -239,7 +258,7 @@ class PyDedupFS(fuse.Fuse):
             # The total number of file serial numbers.
             f_flag = 0, 
             # File system flags. Symbols are defined in the <sys/statvfs.h> header file to refer to bits in this field (see The f_flags field).
-            f_frsize = blocksize, 
+            f_frsize = self.blocksize, 
             # The fundamental file system block size in bytes.
             f_namemax = 4294967295) 
             # The maximum file name length in the file system. Some file systems may return the maximum value that can be stored in an unsigned long to indicate the file system has no maximum file name length. The maximum value that can be stored in an unsigned long is defined in <limits.h> as ULONG_MAX.
@@ -247,18 +266,52 @@ class PyDedupFS(fuse.Fuse):
     def fsinit(self):
         """called after fs initialization"""
         self.logger.debug("PyDedupFS.fsinit()")
-        self.meta_storage = MetaStorage(root="/home/mesznera/pydedupfs")
+        try:
+            options = self.cmdline[0]
+            self.logger.info(options.blocksize)
+            self.base = options.base
+            self.logger.info("Base Directory : %s", self.base)
+            self.blocksize = options.blocksize
+            self.logger.info("Blocksize      : %s", self.blocksize)
+            self.str_hasfunc = options.str_hashfunc
+            self.logger.info("Hash Function  : hashlib.%s", self.hashfunc)
+            self.hashfunc = getattr(hashlib, self.str_hashfunc)
+        except StandardError, exc:
+            self.logger.exception(exc)
+        # initialize meta_storage properly
+        self.meta_storage = MetaStorage(self.base, self.blocksize, self.hashfunc)
+
+        # wrapper to own file class, to pass meta_storage object
+        #class wrapped_file_class(PyDedupFile):
+        #    def __init__(self2, *a, **kw):
+        #        PyDedupFile.__init__(self2, self.meta_storage, *a, **kw)
+        # set own file class
+        #self.file_class = wrapped_file_class
 
     def main(self, *a, **kw):
         # enter endless loop
+        print self.__dict__
+        print sys.argv[-1]
         return fuse.Fuse.main(self, *a, **kw)
 
 
 def main():
+    """main() what else"""
     usage = """PyDedupFS : Python Deduplication Filesystem """+ fuse.Fuse.fusage
     server = PyDedupFS(version="%prog " + fuse.__version__,
                  usage=usage,
                  dash_s_do='setsingle')
+    # put some additional values into parser
+    # TODO should be in -o Parameters to function in fstab entry
+    server.parser.add_option("--base", dest="base", type="str", 
+        help="Directory to store the real data blocks and meta informations")
+    server.parser.add_option("--hashfunc", dest="str_hashfunc", 
+        type="str", default=DEFAULT_HASHFUNC,
+        help="Hashfunction to use <sha1, md5, sha256>, default %default")
+    server.parser.add_option("--blocksize", dest="blocksize", 
+        type="int", default=DEFAULT_BLOCKSIZE,
+        help="Blocksize to use, default %default")
+   
     # does not work with threads
     server.multithreaded = False
     server.parse(values=server, errex=1)
@@ -268,10 +321,19 @@ def main():
 if __name__ == '__main__':
     # add option -f for foreground display an fuse for mountpoint
     # add --base as base directory for real data
-    sys.argv = [sys.argv[0], "-f", "/home/mesznera/fuse"]
-    profile = "PyDedupFS.profile"
-    cProfile.runctx( "main()", globals(), locals(), filename=profile)
-    s = pstats.Stats(profile)
-    s.sort_stats('time')
-    s.print_stats(0.1)
-    os.unlink(profile)
+    # sys.argv = [sys.argv[0], "-f", "--hashfunc=sha1", "--base=/home/mesznera/pydedupfs", "/home/mesznera/fuse"]
+    # TODO: make Profiling a command line switch
+    if "--profile" in sys.argv:
+        # Profiling
+        # dont pass this option to fuse
+        sys.argv.remove("--profile")
+        import cProfile
+        import pstats
+        profile = "PyDedupFS.profile"
+        cProfile.runctx( "main()", globals(), locals(), filename=profile)
+        s = pstats.Stats(profile)
+        s.sort_stats('time')
+        s.print_stats(0.1)
+        os.unlink(profile)
+    else:
+        main()
